User Authentication UI Improvements
- Removed user auth wrapper
- Minor improvements to sign up pop-up
